Reconnect mqtt client
