--- Samples/TestSwitch/CurrentState.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace TestSwitch
-{
-    public class CurrentState
-    {
-        public string OnTime { get; set; } = "11:30";
-        public string OffTime { get; set; } = "12:30";
-    }
-}
