--- Samples/TimeTrigger/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="nanoFramework.CoreLibrary" version="1.17.11" targetFramework="netnano1.0" />
-  <package id="nanoFramework.System.Threading" version="1.1.52" targetFramework="netnano1.0" />
-</packages>
