+++ nanoFramework.HomeAssistant.MqttDiscovery.sln
@@ -3,11 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.14.36414.22
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "TestSwitch", "Samples\TestSwitch\TestSwitch.nfproj", "{35C1667C-A55F-47E4-AC90-685745A904B7}"
-EndProject
-Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "TimeTrigger", "Samples\TimeTrigger\TimeTrigger.nfproj", "{225DE022-8081-430B-A146-C6EAEB94EC7E}"
-EndProject
 Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "nanoFramework.HomeAssistant.MqttDiscovery", "NanoFramework.HomeAssistant.MqttDiscovery\nanoFramework.HomeAssistant.MqttDiscovery.nfproj", "{A91DAAD4-3D96-4543-A598-8F52EF5113C3}"
+EndProject
+Project("{11A8DD76-328B-46DF-9F39-F559912D0360}") = "SwitchSample", "Samples\SwitchSample\SwitchSample.nfproj", "{B334A78A-2B0E-455E-A07F-0E4B8C0E7EFB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,24 +13,18 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{35C1667C-A55F-47E4-AC90-685745A904B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{35C1667C-A55F-47E4-AC90-685745A904B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{35C1667C-A55F-47E4-AC90-685745A904B7}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{35C1667C-A55F-47E4-AC90-685745A904B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{35C1667C-A55F-47E4-AC90-685745A904B7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{35C1667C-A55F-47E4-AC90-685745A904B7}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{225DE022-8081-430B-A146-C6EAEB94EC7E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{225DE022-8081-430B-A146-C6EAEB94EC7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{225DE022-8081-430B-A146-C6EAEB94EC7E}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{225DE022-8081-430B-A146-C6EAEB94EC7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{225DE022-8081-430B-A146-C6EAEB94EC7E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{225DE022-8081-430B-A146-C6EAEB94EC7E}.Release|Any CPU.Deploy.0 = Release|Any CPU
 		{A91DAAD4-3D96-4543-A598-8F52EF5113C3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A91DAAD4-3D96-4543-A598-8F52EF5113C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A91DAAD4-3D96-4543-A598-8F52EF5113C3}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
 		{A91DAAD4-3D96-4543-A598-8F52EF5113C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A91DAAD4-3D96-4543-A598-8F52EF5113C3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A91DAAD4-3D96-4543-A598-8F52EF5113C3}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{B334A78A-2B0E-455E-A07F-0E4B8C0E7EFB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B334A78A-2B0E-455E-A07F-0E4B8C0E7EFB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B334A78A-2B0E-455E-A07F-0E4B8C0E7EFB}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
+		{B334A78A-2B0E-455E-A07F-0E4B8C0E7EFB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B334A78A-2B0E-455E-A07F-0E4B8C0E7EFB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B334A78A-2B0E-455E-A07F-0E4B8C0E7EFB}.Release|Any CPU.Deploy.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
